+++ tomcat/src/main/java/com/techcourse/controller/Controller.java
@@ -2,8 +2,8 @@ package com.techcourse.controller;
 
 import java.io.IOException;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 import com.techcourse.exception.UnsupportedMethodException;
