+++ tomcat/src/main/java/com/techcourse/controller/FrontController.java
@@ -9,11 +9,11 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.MimeType;
-import org.apache.coyote.http11.ResponseBody;
+import org.apache.coyote.http11.response.ResponseBody;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
