+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -3,8 +3,8 @@ package com.techcourse.controller;
 import java.io.IOException;
 import java.util.Map;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
