+++ tomcat/src/main/java/com/techcourse/controller/NotFoundController.java
@@ -7,11 +7,11 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Objects;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.MimeType;
-import org.apache.coyote.http11.ResponseBody;
+import org.apache.coyote.http11.response.ResponseBody;
 
 import com.techcourse.exception.UncheckedServletException;
 import com.techcourse.exception.UnsupportedMethodException;
