+++ tomcat/src/test/java/com/techcourse/controller/FrontControllerTest.java
@@ -9,8 +9,8 @@ import java.io.StringReader;
 import java.net.URL;
 import java.nio.file.Files;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
