+++ tomcat/src/test/java/com/techcourse/controller/LoginControllerTest.java
@@ -8,8 +8,8 @@ import java.io.IOException;
 import java.io.StringReader;
 import java.util.Optional;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
